// src/manifest.d.ts
export type DataSource = DataSourceEnum | DataSourceCustom;
// ...
export interface SAPJSONSchemaForWebApplicationManifestFile {
  /**
   * The resource identifier for the JSON schema to be used. The value of this keyword MUST be a URI [RFC3986] (containing a scheme) and this URI MUST be normalized. The current schema MUST be valid against the meta-schema identified by this URI. If this URI identifies a retrievable resource, that resource SHOULD be of media type "application/schema+json".
   */
  $schema?: string;
  /**
   * Represents Application Descriptor format version. It is managed by schema owner
   */
  _version:
    | "1.1.0"
    | "1.2.0"
    | "1.3.0"
    | "1.4.0"
    | "1.5.0"
    | "1.6.0"
    | "1.7.0"
    | "1.8.0"
    | "1.9.0"
    | "1.10.0"
    | "1.11.0"
    | "1.12.0"
    | "1.13.0"
    | "1.14.0"
    | "1.15.0"
    | "1.16.0"
    | "1.17.0"
    | "1.18.0"
    | "1.19.0"
    | "1.20.0"
    | "1.21.0"
    | "1.22.0"
    | "1.23.0"
    | "1.24.0"
    | "1.25.0"
    | "1.26.0"
    | "1.27.0"
    | "1.28.0"
    | "1.29.0"
    | "1.30.0"
    | "1.31.0"
    | "1.32.0"
    | "1.33.0"
    | "1.34.0"
    | "1.35.0"
    | "1.36.0"
    | "1.37.0"
    | "1.38.0"
    | "1.39.0"
    | "1.40.0"
    | "1.41.0"
    | "1.42.0"
    | "1.43.0"
    | "1.44.0"
    | "1.45.0"
    | "1.46.0"
    | "1.47.0"
    | "1.48.0"
    | "1.49.0"
    | "1.50.0"
    | "1.51.0"
    | "1.52.0"
    | "1.53.0"
    | "1.54.0"
    | "1.55.0"
    | "1.56.0"
    | "1.57.0"
    | "1.58.0"
    | "1.59.0"
    | "1.60.0";
  /**
   * Represents the URL that the developer would prefer the user agent load when the user launches the web application
   */
  start_url?: string;
  "sap.app": JSONSchemaForSAPAPPNamespace;
  "sap.ui": JSONSchemaForSAPUINamespace;
  "sap.ui5"?: JSONSchemaForSAPUI5Namespace;
  "sap.platform.abap"?: JSONSchemaForSAPPLATFORMABAPNamespace;
  "sap.platform.hcp"?: JSONSchemaForSAPPLATFORMHCPNamespace;
  "sap.platform.cf"?: JSONSchemaForSAPPLATFORMCFNamespace;
  "sap.platform.mobilecards"?: JSONSchemaForSAPPLATFORMMOBILECARDSNamespace;
  "sap.fiori"?: JSONSchemaForSAPFIORINamespace;
  "sap.ui.generic.app"?: JSONSchemaForSAPUIGENERICAPPNamespace;
  "sap.fe"?: JSONSchemaForSAPFENamespace;
  "sap.flp"?: JSONSchemaForSAPFLPNamespace;
  "sap.ovp"?: JSONSchemaForSAPOVPNamespace;
  "sap.insights"?: JSONSchemaForSAPInsightsNamespace;
  "sap.wda"?: JSONSchemaForSAPWDANamespace;
  "sap.apf"?: JSONSchemaForSAPAPFNamespace;
  "sap.cloud.portal"?: JSONSchemaForSAPCLOUDPORTALNamespace;
  "sap.gui"?: JSONSchemaForSAPGUINamespace;
  "sap.integration"?: JSONSchemaForSAPINTEGRATIONNamespace;
  "sap.wcf"?: JSONSchemaForSAPWCFNamespace;
  "sap.ui.smartbusiness.app"?: JSONSchemaForSAPUISMARTBUSINESSAPPNamespace;
  "sap.mobile"?: JSONSchemaForSAPMOBILENamespace;
  "sap.copilot"?: JSONSchemaForSAPCOPILOTNamespace;
  "sap.map"?: JSONSchemaForSAPMAPNamespace;
  "sap.url"?: JSONSchemaForSAPURLNamespace;
  "sap.platform.sfsf"?: JSONSchemaForSAPPLATFORMSFSFNamespace;
  "sap.cloud"?: JSONSchemaForSAPCLOUDNamespace;
  "sap.card"?: JSONSchemaForSAPCARDNamespace;
  "sap.package"?: JSONSchemaForSAPPACKAGENamespace;
  "sap.artifact"?: JSONSchemaForSAPARTIFACTNamespace;
  "sap.bpa.task"?: JSONSchemaForSapBpaTaskNamespace;
}
// ...
export interface JSONSchemaForSAPUINamespace {
  /**
   * Represents UI attributes format version. It is managed by namespace owner
   */
  _version?: "1.1.0" | "1.2.0" | "1.3.0" | "1.4.0" | "1.5.0";
  /**
   * Represents UI technology. The possible values are UI5 (default), WDA, NWBC, GUI, URL and WCF
   */
  technology: "UI5" | "WDA" | "NWBC" | "GUI" | "URL" | "WCF";
  /**
   * Represents icons which used in application
   */
  icons?: {
    /**
     * Represents icon of the app
     */
    icon?: string;
    /**
     * Represents ICO file to be used inside the browser and for desktop shortcuts
     */
    favIcon?: string;
    /**
     * Represents 57x57 pixel version for non-retina iPhones
     */
    phone?: string;
    /**
     * Represents 114x114 pixel version for non-retina iPhones
     */
    "phone@2"?: string;
    /**
     * Represents 72x72 pixel version for non-retina iPads
     */
    tablet?: string;
    /**
     * Represents 144x144 pixel version for non-retina iPads
     */
    "tablet@2"?: string;
  };
  /**
   * Represents device types on which application is running. Supported device types are desktop, tablet and phone
   */
  deviceTypes: DeviceType1 & {
    [k: string]: unknown;
  };
  /**
   * The property is Deprecated. Represents array of supported SAP themes such as sap_hcb, sap_bluecrystal
   */
  supportedThemes?: string[];
  /**
   * Indicates whether app should run in full screen mode: possible values: true or false
   */
  fullWidth?: boolean;
  [k: string]: unknown;
}
// ...
export interface Model {
  /**
   * Represents model class name
   */
  type?: string;
  /**
   * Represents string of key/alias from sap.app dataSources to reference an existing data source
   */
  dataSource?: string;
  /**
   * Represents URI for the model
   */
  uri?: string;
  /**
   * Indicates that the model will be immediately created after the manifest is loaded by Component Factory and before the Component instance is created
   */
  preload?: boolean;
  settings?: Ui5Setting;
}
// ...
export interface JSONSchemaForSAPPLATFORMABAPNamespace {
  /**
   * Represents attributes format version. It is managed by namespace owner
   */
  _version?: "1.1.0" | "1.2.0";
  /**
   * Represents the uri of the app in the ABAP system
   */
  uri?: string;
  /**
   * Represents the alternative uri of the app in the ABAP system for starting the application
   */
  uriNwbc?: string;
}
/**
 * Represents HANA Cloud Platform platform specific attributes
 */
export interface JSONSchemaForSAPPLATFORMHCPNamespace {
  /**
   * Represents attributes format version. It is managed by namespace owner
   */
  _version?: "1.1.0" | "1.2.0" | "1.3.0";
  /**
   * Represents the uri of the app in the HANA Cloud Platform
   */
  uri?: string;
  /**
   * Represents the alternative uri of the app in the ABAP system for starting the application
   */
  uriNwbc?: string;
  /**
   * Represents the provider account of the HTML5 application
   */
  providerAccount?: string;
  /**
   * Represents the HTML5 application name
   */
  appName?: string;
  /**
   * Represents the version of the HTML5 application
   */
  appVersion?: string;
  /**
   * Indicates that HCP application is multi-version enabled
   */
  multiVersionApp?: boolean;
}
/**
 * Represents CF(Cloud Foundry) platform specific attributes
 */
export interface JSONSchemaForSAPPLATFORMCFNamespace {
  /**
   * Represents attributes format version. It is managed by namespace owner
   */
  _version?: "1.1.0" | "1.2.0";
  /**
   * Represents the authorization scope of the application
   */
  oAuthScopes?: string[];
  [k: string]: unknown;
}
/**
 * Represents Mobile Cards platform specific attributes
 */
export interface JSONSchemaForSAPPLATFORMMOBILECARDSNamespace {
  /**
   * Represents attributes format version. It is managed by namespace owner
   */
  _version?: "1.0.0";
  /**
   * Represents the compatibility of this app with the Mobile Cards platform.
   */
  compatible?: boolean;
  [k: string]: unknown;
}
/**
 * Represents SAP Fiori  specific attributes
 */
export interface JSONSchemaForSAPFIORINamespace {
  /**
   * Represents attributes format version. It is managed by namespace owner
   */
  _version?: "1.1.0" | "1.2.0" | "1.3.0";
  /**
   * Represents array of registration ids, i.e. for Fiori apps fiori id(s)
   */
  registrationIds: string[];
  /**
   * Represents architecture type of an application. The supported types are transactional or analytical or factsheet or reusecomponent or fpmwebdynpro or designstudio
   */
  archeType: "transactional" | "analytical" | "factsheet" | "reusecomponent" | "fpmwebdynpro" | "designstudio";
  /**
   * Indicator that app is an abstract (generic) app which may not be used directly, but needs to be specialized in the SAP Fiori launchpad content
   */
  abstract?: boolean;
  /**
   * Represents the release status for the developer adaptation in the cloud (relevant for SAP internal only). The supported types are released, deprecated, obsolete, no value means not released
   */
  cloudDevAdaptationStatus?: "released" | "deprecated" | "obsolete";
}
/**
 * Represents GENERIC APP specific attributes
 */
export interface JSONSchemaForSAPUIGENERICAPPNamespace {
  _version?: "1.1.0" | "1.2.0" | "1.3.0" | "1.4.0" | "1.5.0" | "1.6.0";
  settings?: SettingDef;
  /**
   * Represents one ore more pages of an application. UI5 routing is created from the definitions in this section
   */
  pages?:
    | PagesArray[]
    | {
        [k: string]: PagesMap;
      };
}
// ...
export interface JSONSchemaForSAPFENamespace {
  [k: string]: unknown;
}
/**
 * Represents FLP specific attributes
 */
export interface JSONSchemaForSAPFLPNamespace {
  /**
   * Represents attributes format version. It is managed by namespace owner
   */
  _version?: "1.1.0" | "1.2.0" | "1.3.0";
  /**
   * Represents size of the tile
   */
  tileSize?: "1x1" | "1x2";
  /**
   * Represents the type of FLP entry. It must be 'application' or 'tile' or 'plugin'
   */
  type: "application" | "tile" | "plugin";
  /**
   * Represents configuration parameters of the FLP entry
   */
  config?: {
    /**
     * Represents the configuration key and values
     *
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^[a-zA-Z0-9\_\.\-]*$".
     */
    [k: string]: {
      [k: string]: unknown;
    };
  };
  /**
   * Represents the original tile and target mapping which resulted in this app
   */
  origin?: {
    /**
     * Represents the original tile which resulted in this app
     */
    tileId?: string;
    /**
     * Represents the original target mapping which resulted in this app
     */
    targetMappingId?: string;
  };
}
// ...
export interface JSONSchemaForSAPInsightsNamespace {
  _version?: "1.0.0" | "1.1.0" | "1.2.0" | "1.3.0" | "1.4.0" | "1.5.0";
  /**
   * Represents mandatory unique app identifier of the app containing self manifest.
   */
  parentAppId: string;
  /**
   * Represents the entity to use as global filter in the filter bar control
   */
  filterEntitySet?: string;
  /**
   * Represents the template name, from where manifest is generated
   */
  templateName?: string;
  /**
   * Represents the environment type of Insights card
   */
  cardType: "DT" | "RT";
  /**
   * Represents the state of original or copied card
   */
  isDtCardCopy?: boolean;
  /**
   * Represents the deleted state for a user
   */
  isDeletedForUser?: boolean;
  /**
   * Represents the visibility type of Insights card
   */
  visible?: boolean;
  /**
   * Deprecated: Represents the display order of Insights card
   */
  rank?: number;
  /**
   * Represents UI5 and card generator middleware version.
   */
  versions?: {
    /**
     * Represents the version of UI5 used to generate the card.
     */
    ui5?: string;
    /**
     * Represents the version of design time card generator
     */
    dtMiddleware?: string;
    [k: string]: unknown;
  };
  /**
   * New property representing the display order of Insights card
   */
  ranking?: string;
  /**
   * Array of all supported chart type allowed for transformation.
   */
  allowedChartTypes?: (
    | "bar"
    | "column"
    | "line"
    | "bullet"
    | "vertical_bullet"
    | "100_stacked_bar"
    | "100_stacked_column"
    | "waterfall"
    | "horizontal_waterfall"
    | "area"
    | "radar"
    | "combination"
    | "stacked_bar"
    | "stacked_column"
    | "stacked_combination"
    | "horizontal_stacked_combination"
    | "pie"
    | "donut"
    | "dual_bar"
    | "dual_column"
    | "dual_line"
    | "dual_stacked_bar"
    | "dual_stacked_column"
    | "dual_combination"
    | "dual_horizontal_combination"
    | "dual_stacked_combination"
    | "dual_horizontal_stacked_combination"
    | "100_dual_stacked_bar"
    | "100_dual_stacked_column"
    | "scatter"
    | "heatmap"
    | "timeseries_column"
    | "timeseries_line"
    | "timeseries_scatter"
    | "timeseries_stacked_column"
    | "timeseries_100_stacked_column"
    | "timeseries_bullet"
    | "timeseries_waterfall"
    | "timeseries_stacked_combination"
  )[];
}